Reduce maximum value for redstone core price to minimize potential overflow risk.

+++ contracts/oracles/adaptors/redstone/BaseRedstoneCoreAdaptor.sol
@@ -108,10 +108,11 @@ abstract contract BaseRedstoneCoreAdaptor is BaseOracleAdaptor {
             adaptorData = adaptorDataNonUSD[asset];
         }
 
-        // We use a maximum buffered price of 2^216 - 1 since redstone core reports pricing 
-        // in 8 decimal format, requiring multiplication by 10e10 to standardize to 18 decimal
-        // format, which overflows soon after 2^216 - 1.
-        adaptorData.max = type(uint216).max;
+        // We use a maximum buffered price of 2^192 - 1 since redstone core
+        // reports pricing in 8 decimal format, requiring multiplication by
+        // 10e10 to standardize to 18 decimal format, which could overflow 
+        // when trying to save the final value into an uint240.
+        adaptorData.max = type(uint192).max;
         adaptorData.symbolHash = symbolHash;
         adaptorData.isConfigured = true;
         isSupportedAsset[asset] = true;
